Fixed gui

+++ src/main/java/hackstyle/scripts/actions/SleepAction.java
@@ -15,7 +15,7 @@ public class SleepAction implements ScriptAction {
     @Override
     public void act(State state) {
         try {
-            Thread.sleep(Long.parseLong(variable.get()));
+            Thread.sleep(Integer.parseInt(variable.get()));
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
